OS X nightly build: Skip missing 32-bit device adapters.
(Just to make it easier to test other aspects of the script without
installing every camera framework.)

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@13627 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ buildscripts/nightly/nightlybuild_OSX_mm.sh
@@ -120,8 +120,10 @@ for file in DeviceAdapters/PVCAM/.libs/libmmgr_dal_PVCAM \
             DeviceAdapters/ScionCam/.libs/libmmgr_dal_ScionCam \
             DeviceAdapters/Spot/.libs/libmmgr_dal_Spot 
 do
-   lipo -extract i386 -output $file.i386 $file
-   mv $file.i386 $file
+   if [ -f $file ]; then
+      lipo -extract i386 -output $file.i386 $file
+      mv $file.i386 $file
+   fi
 done
